refactor: edit tools
